chore: update creating model with all tags
